% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{adj}
\alias{adj}
\title{get adj}
\usage{
adj(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\description{
\code{adj} returns the elements of adj, of type either  in an object
}
