% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ALADIN_Jyvaskyla}
\alias{ALADIN_Jyvaskyla}
\title{ALADIN5.3 EURO-CORDEX 0.11 deg resolution.}
\format{A data frame with 54819 rows and 3 variables:
\describe{
  \item{date}{time stamps for each day}
  \item{tas}{daily mean temperature in Celcius}
  \item{pr}{daily precipitation in mm/d}
  ...
}}
\source{
\url{https://esgf.llnl.gov/}
}
\usage{
ALADIN_Jyvaskyla
}
\description{
A dataset containing the time series of 2m daily mean temperature and daily precipitation
simulated with CNRM-ALADIN5.3 at 0.11 degree resolution using RCP4.5 emission scenario forcings
in a grid box nearest to Jyväskylä airport (25.68 E, 62.40 N).
}
\keyword{datasets}
