% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasco.R
\name{biasco}
\alias{biasco}
\title{A wrapper function for univariate bias correction and delta changes methods.}
\usage{
biasco(obs.in, ctrl.in, scen.in, type = "abs", method, ...)
}
\arguments{
\item{obs.in}{A vector containing the reference data}

\item{ctrl.in}{A vector of simulated data used for calibration}

\item{scen.in}{A vector of data to be adjusted}

\item{type}{String defining the type of the input. If the variable in on absolute scale, the}

\item{method}{Method to be used to adjust the}

\item{...}{Additional parameters to used in bias correction methods}
}
\value{
An object of type \code{BiascoTimeSeries},
  which contains the adjusted data and parameter values related to a particular method
}
\description{
\code{biasco} returns bias corrected data as and object of type \code{BiascoTimeSeries}
}
\examples{
TODO
}
\seealso{
\code{\link{biasco2d}} for two-dimensional bias correction of temperature and precipitation
}
