% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{ctrl}
\alias{ctrl}
\title{get ctrl}
\usage{
ctrl(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\description{
\code{ctrl} returns the content of ctrl,
  which can be either type TimeSeries or TimeSeriesTP and can also be part of other object.
}
