% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{.adjPT}
\alias{.adjPT}
\title{.adjPT}
\usage{
.adjPT(obs, ctrl, scen, fit.skew, eps = 1e-15)
}
\arguments{
\item{obs}{Observed time series.}

\item{ctrl}{Control period time series.}

\item{scen}{Scenario period time series.}

\item{eps}{Small value used to handle zero and unity probabilities.}
}
\value{
Adjusted time series of temperature and precipitation.
}
\description{
Conditional adjustment of temperature with respect to precipitation (internal).
}
\keyword{internal}
