% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{.adjTP}
\alias{.adjTP}
\title{.adjTP}
\usage{
.adjTP(obs, ctrl, scen, fit.skew, eps = 1e-15)
}
\arguments{
\item{obs}{A list containing the observed time series of temperature and precipitation.}

\item{ctrl}{A list containing the simulated time series of temperature and precipitation in the control period.}

\item{scen}{A list containing the simulated time series of temperature and precipitation in the scenario period.}

\item{eps}{Small value used to handle zero and unity probabilities.}
}
\value{
Adjusted time series of temperature and precipitation.
}
\description{
Conditional adjustment of precipitation with respect to temperature (internal).
}
\keyword{internal}
