% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{.adjustEs}
\alias{.adjustEs}
\title{Engen-Skaugen algorithm (internal)}
\usage{
.adjustEs(data, mean.target, sd.target, tol = 0.001, max.iter = 100)
}
\arguments{
\item{data}{Data to be adjusted}

\item{mean.target}{Traget value for the mean}

\item{sd.target}{Target value for standard deviation}

\item{tol}{Tolerance for the convergence}

\item{max.iter}{Max. number of iterations, after which the execution is stopped}
}
\value{
Adjusted time series for the scenario period.
}
\description{
Adjust the cv interatively using the Engen-Skaugen algorithm.
}
\keyword{internal}
