% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{.BcMean}
\alias{.BcMean}
\title{Bias correction of the mean}
\usage{
.BcMean(.Object, ...)
}
\arguments{
\item{.Object}{An object}

\item{...}{Additional parameters required by a specific implementation}
}
\value{
Bias corrected time series
}
\description{
A generic for bias correction of mean
}
\keyword{internal}
