% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{.BcMeanSd2}
\alias{.BcMeanSd2}
\title{Bias correction of the mean and standard deviation}
\usage{
.BcMeanSd2(.Object, ...)
}
\arguments{
\item{.Object}{An object}

\item{...}{Additional parameters required by a specific implementation}
}
\description{
A generic for bias correction of mean and standard deviation
}
\keyword{internal}
