% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{.BcMeanSdSkew}
\alias{.BcMeanSdSkew}
\title{Bias correction of the mean, standard deviation and skewness}
\usage{
.BcMeanSdSkew(.Object, ...)
}
\arguments{
\item{.Object}{An object}

\item{...}{Additional parameters required by a specific implementation}
}
\value{
Bias corrected time series
}
\description{
A generic for bias correction of mean, stardard deviation and skewness
}
\examples{
Needs to be written
}
\keyword{internal}
