% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{.BcQmEmpir}
\alias{.BcQmEmpir}
\title{Bias correction using empirical quantile mapping}
\usage{
.BcQmEmpir(.Object, ...)
}
\arguments{
\item{.Object}{An object}

\item{...}{Additional parameters required by a specific implementation}
}
\value{
Bias corrected time series
}
\description{
A generic for bias correction using empirical quantile mapping
}
\keyword{internal}
