% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{.BcQmParam}
\alias{.BcQmParam}
\title{Bias correction using parametric quantile mapping}
\usage{
.BcQmParam(.Object, ...)
}
\arguments{
\item{.Object}{An object}

\item{...}{Additional parameters required by a specific implementation}
}
\value{
Bias corrected time series
}
\description{
A generic for bias correction using parametric quantile mapping
}
\keyword{internal}
