% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{.DcMean}
\alias{.DcMean}
\title{Mean delta change}
\usage{
.DcMean(.Object, ...)
}
\arguments{
\item{.Object}{An object}

\item{...}{Additional parameters required by a specific implementation}
}
\value{
Delta change adjusted time series
}
\description{
A generic for mean delta change scaling
}
\keyword{internal}
