% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{.DcMeanSdSkew}
\alias{.DcMeanSdSkew}
\title{Delta change scaling of mean, standard deviation and skewness}
\usage{
.DcMeanSdSkew(.Object, ...)
}
\arguments{
\item{.Object}{An object}

\item{...}{Additional parameters required by a specific implementation}
}
\value{
Delta change adjusted time series
}
\description{
A generic for delta change scaling of mean and standard deviation.
}
\examples{
Needs to be written
}
\keyword{internal}
