% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{.DcQmEmpir}
\alias{.DcQmEmpir}
\title{Empirical quantile mapping applied in the delta change form}
\usage{
.DcQmEmpir(.Object, ...)
}
\arguments{
\item{.Object}{An object}

\item{...}{Additional parameters required by a specific implementation}
}
\value{
Delta change adjusted time series
}
\description{
A generic for quantile mapping applied in the delta change mode.
}
\examples{
Needs to be written
}
\keyword{internal}
