% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{.DcQmParam}
\alias{.DcQmParam}
\title{Parametric quantile mapping applied in the delta change form}
\usage{
.DcQmParam(.Object, ...)
}
\arguments{
\item{.Object}{An object}

\item{...}{Additional parameters required by a specific implementation}
}
\value{
Delta change adjusted time series
}
\description{
A generic for quantile mapping applied in the delta change mode.
}
\keyword{internal}
