% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{.fitMarginal}
\alias{.fitMarginal}
\title{.fitMarginal}
\usage{
.fitMarginal(data, type)
}
\arguments{
\item{data}{Data used to fit the selected distribution}

\item{type}{Type of distribution to be fitted}
}
\value{
Parameters of the fitted distribution
}
\description{
Fit parametric marginal distribution to the data (internal)
}
\details{
#Two options are possible:
-gaussian marginal for temperature
-gamma marginal for precipitation
}
\keyword{internal}
