% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{.iterAb}
\alias{.iterAb}
\title{.iterAb}
\usage{
.iterAb(
  data,
  mean.target,
  sd.target,
  b.lim = c(0.001, 5),
  tol = 0.001,
  max.iter = 100
)
}
\arguments{
\item{data}{Input data}

\item{mean.target}{Target values for the mean}

\item{sd.target}{Target value for standard deviations}

\item{b.lim}{Limits for the exponent value}

\item{tol}{Tolerance for the convergence}

\item{max.iter}{Max. number of iterations. Used to stop the execution,
if the algorithm does not converge to the target values.}
}
\value{
Adjusted time series for the scenario period.
}
\description{
Power transformation algorithm (internal)
}
\details{
Find iteratively the correct exponent in the power transformation of a ratio variable.
}
\keyword{internal}
