% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{.JBC}
\alias{.JBC}
\title{Joint bias correction of temperature and precipitation}
\usage{
.JBC(.Object, cond, threshold, ...)
}
\arguments{
\item{.Object}{An object}

\item{cond}{character flag indicating the order of conditioning. If value "T" is given temperature is bias corrected first,
while "P" is selected the adjustment of temperature is conditioned on precipitation.}

\item{threshold}{The value used to divide precipitation time series to wet and dry days.}

\item{...}{Additional parameters required by the specific implementation of \code{JBC}.}
}
\value{
Bias corrected time series of both temperature and precipitation.
}
\description{
A generic for joint bias correction of temperature and precipitation using the algorithm by Li et al. (2014)
}
\keyword{internal}
