% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{.minimizeMSE}
\alias{.minimizeMSE}
\title{.minimizeMSE}
\usage{
.minimizeMSE(ref, cal, val)
}
\arguments{
\item{ref}{Reference data}

\item{cal}{Calibration data}

\item{val}{Scenario data}
}
\value{
Adjusted calibration and scenario data.
}
\description{
Minimize MSE between the reference and simulated time series (internal)
}
\keyword{internal}
