% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{.ppRain}
\alias{.ppRain}
\title{Pre-process precipitation data for parametric quantile mapping.}
\usage{
.ppRain(ref, adj, ref.threshold = 0.1)
}
\arguments{
\item{ref}{Reference data}

\item{adj}{Data to be adjusted}

\item{ref.threshold}{Threshold for wet-day values.}
}
\value{
A list containing the number of dry days both in the reference and calibration data sets.
}
\description{
Pre-process precipitation data for parametric quantile mapping.
}
\keyword{internal}
