% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{.quantMapP}
\alias{.quantMapP}
\title{Quantile mapping algorithm for precipitation (internal)}
\usage{
.quantMapP(Scen, Ctrl, Obs)
}
\arguments{
\item{Scen}{scenario period time series}

\item{Ctrl}{control period time series}

\item{Obs}{control period observations}
}
\value{
Adjusted time series for the Scenario period
}
\description{
Determine the future temperatures (ScenCor) that
}
\keyword{internal}
