% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{.quantMapT}
\alias{.quantMapT}
\title{.quantMapT}
\usage{
.quantMapT(Scen, Ctrl, Obs)
}
\arguments{
\item{Scen}{scenario period time series}

\item{Ctrl}{control period time series}

\item{Obs}{control period observations}
}
\value{
Bias adjusted time series
}
\description{
Quantile mapping algorithm (internal)
}
\details{
Determine the future temperatures (ScenCor) that
}
\keyword{internal}
