% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{.zeros}
\alias{.zeros}
\title{Handle zero values (internal)}
\usage{
.zeros(Data, eps)
}
\arguments{
\item{Data}{Time series to be adjusted}

\item{eps}{small value used to adjust the zero values}
}
\value{
adjusted time series
}
\description{
Handle zero values (internal)
}
\keyword{internal}
