% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbsCorrectionObject-class.R
\name{initialize,AbsCorrection-method}
\alias{initialize,AbsCorrection-method}
\title{initialize}
\usage{
\S4method{initialize}{AbsCorrection}(
  .Object,
  ...,
  obs = new("TimeSeries"),
  ctrl = new("TimeSeries"),
  scen = new("TimeSeries")
)
}
\description{
initialize
}
