% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{method}
\alias{method}
\title{get method}
\usage{
method(object, value, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\description{
\code{method} returns the string stored in the slot method.
}
