% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{nvar<-}
\alias{nvar<-}
\title{Set nvar value}
\usage{
nvar(object) <- value
}
\arguments{
\item{object}{An object}

\item{value}{Value to be set slot nvar}
}
\value{
new value for nvar
}
\description{
Set nvar value
}
\keyword{internal}
