% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbsCorrectionObject-class.R,
%   R/JointCorrectionObject-class.R
\name{show,AbsCorrection-method}
\alias{show,AbsCorrection-method}
\alias{show,JointCorrection-method}
\title{show}
\usage{
\S4method{show}{AbsCorrection}(object)

\S4method{show}{JointCorrection}(object)
}
\description{
show
}
