% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station_Jyvaskyla}
\alias{station_Jyvaskyla}
\title{Temperature and precipitation in Jyv채skyl채 between 1951-2017.}
\format{A data frame with 24139 rows and 3 variables:
\describe{
  \item{date}{time stamps for each day}
  \item{tas}{daily mean temperature in Celcius}
  \item{pr}{daily precipitation in mm/d}
  ...
}}
\source{
\url{https://climexp.knmi.nl/}
}
\usage{
station_Jyvaskyla
}
\description{
A data.frame containing the daily time series of daily mean temperature and daily precipitation
measures in the Jyv채skyl채 airport (25.68 E, 62.40 N) between 1951-2017.
Missing values have been replaced with previous day's values.
}
\keyword{datasets}
