% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{type<-}
\alias{type<-}
\title{Set a new value for slot type}
\usage{
type(object) <- value
}
\arguments{
\item{object}{An object}

\item{value}{Value given for \code{slot}}
}
\description{
Set a new value for slot type
}
