% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{varsC}
\alias{varsC}
\title{Synthetic data for demonstrating the use of BCUH.}
\format{
A data frame with 10000 rows and 2 variables:
\describe{
  \item{temp}{Synthetic daily mean temperature in Celcius}
  \item{prec}{Synthetic daily precipitation in mm/d}
  ...
}
}
\usage{
varsC
}
\description{
A data.frame synthetic time series for daily mean temperature and daily precipitation.
These data act as simulated data in the demonstration.
}
\keyword{datasets}
