% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasco.R
\name{biasco}
\alias{biasco}
\title{A wrapper function for univariate bias correction and delta changes methods.}
\usage{
biasco(obs.in, ctrl.in, scen.in, type = "abs", method = "M1", ...)
}
\arguments{
\item{obs.in}{A vector containing the reference data.}

\item{ctrl.in}{A vector of simulated data used to calibrate the method}

\item{scen.in}{A vector of data for the scenario period simulation}

\item{type}{String defining the type of adjustment. Currently \code{type} recognizes two values:
\itemize{
  \item{abs: Absolute scale is used. This approach is typically applied measured in the absolute scale (e.g temperature)}
  \item{ratio: The adjustment is made in ratio format. This approach is typically applied to non-continuous variables such as precipitation.}
}}

\item{method}{Method to be used to bias adjust \code{scen.in}. The following options are available:}

\item{...}{Additional parameters used to handle the behavior of specific bias correction methods.}
}
\value{
An object of type \code{BiascoTimeSeries},
  which contains the adjusted data and parameter values related to a particular method.
}
\description{
\code{biasco} returns bias corrected data as an object of type \code{BiascoTimeSeries}. 
Several methods of different complexity are available in the package (see details).
}
\details{
The following univariate methods have currently been implemented in this package:
 \itemize{
 \item{M1: Simple delta change scaling of time mean}
 \item{M2: Delta change scaling on mean and standard deviation. If \code{type} = ratio, then Engen-Skaugen algorithm is used.}{}
 \item{M3: Delta change scaling on mean, standard deviation and skewness.  If \code{type} = ratio, M3 corresponds to delta change scaling of mean and standard deviation using.}
 \item{M4: Non-parametric quantile mapping applied in the delta change mode}
 \item{M5: Parametric quantile mapping applied in the delta change mode}
 \item{M6: As M1 but applied in the bias correction mode}
 \item{M7: As M2 but applied in the bias correction mode}
 \item{M8: As M3 but applied in the bias correction mode}
 \item{M9: As M4 but applied in the bias correction mode}
 \item{M10: As M5 but applied in the bias correction mode}
 }
}
\examples{

library(lubridate)
library(BCUH)

data("obs_data")
data("model_data")

#Create objects which contain the bias corrected data
bc6 <- biasco(varsO[,"temp"], varsC[,"temp"], varsC[,"temp"], type = "abs", method = "M6")
bc9 <- biasco(varsO[,"temp"], varsC[,"temp"], varsC[,"temp"], type = "abs", method = "M9")
#Visualise the results
plot(quantile(dat(adj(bc6)), seq(0,1,0.01)), type = "l",
     main="Quantile plot", xlab = "\%", ylab = "Celcius", ylim = c(10,30))
lines(quantile(dat(adj(bc9)), seq(0,1,0.01)), col = "black",lty=2)
lines(quantile(varsO[,"temp"], seq(0,1,0.01)), col = "red")
lines(quantile(varsC[,"temp"], seq(0,1,0.01)), col = "blue")
lines(quantile(varsC[,"temp"], seq(0,1,0.01)), col = "blue", lty = 2)
legend("topleft", c("M6","M9","Obs","Ctrl","Scen"),
       col = c("black","black","red","blue","blue"), lty = c(1,2,1,1,2))



}
\seealso{
\code{\link{biasco2d}} for two-dimensional bias correction of temperature and precipitation
}
