% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasco.R
\name{biasco2D}
\alias{biasco2D}
\title{Joint bias correction of temperature and precipitation.}
\usage{
biasco2D(obs.in, ctrl.in, scen.in, names = NULL, cond = "P",
  threshold = 0.1, ...)
}
\arguments{
\item{obs.in}{Either a 2-column array, matrix or data.frame. of observed temperature and precipitation}

\item{ctrl.in}{Data used to calibrate the method given in the format as \code{obs.in}}

\item{scen.in}{Data for the scenario period simulation given in the format as \code{obs.in}}

\item{cond}{The conditioning order of temperature and precipitation. If \code{cond} = "P" then the precipitation marginal
is adjusted first, while the adjustment of temperature is made conditionally on precipitation}

\item{threshold}{Threshold for wet day precipitation. Everything below \code{threshold} is handled as zero precipitation}

\item{...}{Method-specific parameters to be implemented}
}
\value{
An object of type BiascoTimeSeriesPT which contains a matrix of bias adjusted temperature and precipitation and also
additional information on the used parameter values
}
\description{
A wrapper function for a copula-based bias correction method of temperature and precipitation
}
\examples{

library(lubridate)
library(BCUH)
#Load data
data("station_Jyvaskyla")
data("ALADIN_Jyvaskyla")

#Extract data for January
ctrl <- c(1971,2000)
scen <- c(2071,2100)
ind.obs <- which(month(station_Jyvaskyla$date) == 1 &
                  year(station_Jyvaskyla$date) \%in\% seq(ctrl[1],ctrl[2]))
ind.ctrl <- which(month(ALADIN_Jyvaskyla$date) == 1 &
                   year(ALADIN_Jyvaskyla$date) \%in\% seq(ctrl[1],ctrl[2]))
ind.scen <- which(month(ALADIN_Jyvaskyla$date) == 1 &
                   year(ALADIN_Jyvaskyla$date) \%in\% seq(scen[1],scen[2]))
obs.ctrl <- station_Jyvaskyla[ind.obs,2:3]
rcm.ctrl <- ALADIN_Jyvaskyla[ind.ctrl,2:3]
rcm.scen <- ALADIN_Jyvaskyla[ind.scen,2:3]
biasco2d.object <- biasco2D(obs.ctrl,rcm.ctrl,rcm.scen, names = c("tas", "pr"))

#Visualise the results:

par(mfrow=c(1,2))
plot(obs.ctrl)
points(rcm.ctrl,col="red")
legend("topleft",c("Obs","Ctrl"),col=c("black","red"),pch=c(1,1))
plot(dat(adj(biasco2d.object)))
points(rcm.scen,col="red")
legend("topleft",c("Adj","Scen"),col=c("black","red"),pch=c(1,1))

}
