% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biasco.R
\name{biasco2D}
\alias{biasco2D}
\title{Joint bias correction of temperature and precipitation.}
\usage{
biasco2D(
  obs.in,
  ctrl.in,
  scen.in,
  names = NULL,
  cond = "P",
  separate = F,
  threshold = 0.1,
  jittering = F,
  fit.skew = F,
  ...
)
}
\arguments{
\item{obs.in}{Either a 2-column array, matrix or data.frame. of observed temperature and precipitation}

\item{ctrl.in}{Data used to calibrate the method given in the format as \code{obs.in}}

\item{scen.in}{Data for the scenario period simulation given in the format as \code{obs.in}}

\item{names}{names of the variables at different columns}

\item{cond}{The conditioning order of temperature and precipitation. If \code{cond} = "P" then the precipitation marginal
is adjusted first, while the adjustment of temperature is made conditionally on precipitation}

\item{threshold}{Threshold for wet day precipitation. Everything below \code{threshold} is handled as zero precipitation}

\item{...}{Method-specific parameters to be implemented}
}
\value{
An object of type BiascoTimeSeriesPT which contains a matrix of bias adjusted temperature and precipitation and also
additional information on the used parameter values
}
\description{
A wrapper function for a copula-based bias correction method of temperature and precipitation
}
\examples{

library(lubridate)
library(BCUH)
library(copula)

data("obs_data")
data("model_data")
biasco2d.object <- biasco2D(varsO,varsC,varsC,cond="P",
                            names = c("temp", "prec"), 
                            threshold=0.1)

#Visualise the results:

par(mfrow=c(1,2))
plot(varsO)
points(varsC,col="red")
legend("topleft",c("Obs","Ctrl"),col=c("black","red"),pch=c(1,1))

plot(dat(adj(biasco2d.object)))
points(varsO,col="red")
legend("topleft",c("Adj","Scen"),col=c("black","red"),pch=c(1,1))
}
