% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R,
%   R/AbsCorrectionObject-class.R, R/JointCorrectionObject-class.R
\name{ctrl}
\alias{ctrl}
\alias{ctrl,AbsCorrection-method}
\alias{ctrl,JointCorrection-method}
\title{get ctrl}
\usage{
ctrl(object, ...)

\S4method{ctrl}{AbsCorrection}(object)

\S4method{ctrl}{JointCorrection}(object)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\description{
\code{ctrl} returns the content of ctrl,
  which can be either type TimeSeries or TimeSeriesTP and can also be part of other object.
}
