% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{.fitMarginal}
\alias{.fitMarginal}
\title{Fit parametric marginal distribution to the data (internal)}
\usage{
.fitMarginal(data, type)
}
\arguments{
\item{data}{Data used to fit the selected distribution}

\item{type}{Type of distribution to be fitted}
}
\value{
Parameters of the fitted distribution
}
\description{
#Two options are possible:
}
\keyword{internal}
