% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxfunctions.R
\name{.iterSkewness}
\alias{.iterSkewness}
\title{.iterSkewness}
\usage{
.iterSkewness(
  source,
  skew.target,
  nseq = 1,
  tol = 0.001,
  a.lim = c(0.5, 9),
  iter.max = 100
)
}
\arguments{
\item{skew.target}{target skewness}

\item{nseq}{data set is divided into nseq chunks on which the skewness is adjusted separately}

\item{tol}{threshold for iteration}

\item{a.lim}{limits for the exponent used to adjust the skewness}

\item{iter.max}{max. number of iterations}

\item{data}{data to be adjusted}
}
\value{
Bias adjusted time series
}
\description{
Find correct skewness iteratively (internal)
}
\keyword{internal}
