% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R
\name{obs}
\alias{obs}
\title{get obs}
\usage{
obs(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\description{
\code{obs} returns the content of obs,
  which can be either type TimeSeries or TimeSeriesTP and can also be part of other object.
}
