% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/AbsCorrectionObject-class.R
\name{obs}
\alias{obs}
\alias{obs,AbsCorrection-method}
\title{get obs}
\usage{
obs(object, ...)

\S4method{obs}{AbsCorrection}(object)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\description{
\code{obs} returns the content of obs,
  which can be either type TimeSeries or TimeSeriesTP and can also be part of other object.
}
