% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericMethods.R, R/AbsCorrectionObject-class.R
\name{scen}
\alias{scen}
\alias{scen,AbsCorrection-method}
\title{scen}
\usage{
scen(object, ...)

\S4method{scen}{AbsCorrection}(object)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\description{
\code{scen} returns the content of scen,
  which can be either type TimeSeries or TimeSeriesTP and can also be part of other object.
}
